Reorganize tasks into folder. Fix views function. Use put
