+++ packages/lib-classifier/src/store/tasks/Task.js
@@ -1,0 +1,7 @@
+import { types } from 'mobx-state-tree'
+
+const Task = types.model('Task', {
+  taskKey: types.identifier
+})
+
+export default Task
